requirements.txt
# Telegram (включает httpx)
python-telegram-bot[http2]==20.7

# Database
sqlalchemy==2.0.25

# Environment
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Development
pytest==7.4.4
black==24.1.1
flake8==7.0.0
